revert: improve forward-compatibility of exposed config properties (#172)
The revert is necessary because of issues when injecting environment variables as string in places where other types of values are expected (e.g., booleans, ints, etc).

Therefore, it is necessary to specify the types even tho it implies more risks of drifts. However, such risk is in fact quite low is will most probably not have big consequences.

This reverts commit 1e8a7f6396cfd3e4c1ad3bbb3de4325f854e8116.