revert: improve forward-compatibility of exposed config properties (#174)
The revert of #172 is necessary because of issues when injecting environment variables as string in places where other types of values are expected (e.g., booleans, ints, etc).

On top of that, PaperMC seems very bad at handling misconfigurations since specifying `'true'` (String form of a boolean) instead of `true` (a unambiguous boolean value in YAML) will lead to the property to be simply ignored without any warning.
But the fact is, even if PaperMC would detect and raise proper warnings/errors in such circunstances, it would not change the fact values inferred from environment variables are incorrect for most properties, which is unacceptable.

Therefore, it is necessary to specify the types even tho it implies more risks of drifts. However, such risk is in fact quite low is will most probably not have big consequences.

This reverts commit 1e8a7f6396cfd3e4c1ad3bbb3de4325f854e8116.