build: fix pip-compile Docker wrapper on Linux systems (#256)
The permission issue is not a concern on Windows, hence why the issue was not spotted at first.

The solution consists on building the Docker image with both the UID and GID of the user running the script.
This is acceptable since the Docker image is systematically built before being executed.